# Backend/api_documentation.yaml
openapi: 3.1.0
info:
  title: CivicResolve AI
  version: 0.1.0
paths:
...
  /api/citizen/dash:
    get:
      tags:
      - citizen
      summary: Citizen Dashboard
      operationId: citizen_dashboard_api_citizen_dash_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitizenDashboardResponse'
      security:
      - HTTPBearer: []
  /api/citizen/bills:
    get:
      tags:
      - citizen
      summary: Citizen Bills List
      operationId: citizen_bills_list_api_citizen_bills_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/UtilityBillSchema'
                title: Response Citizen Bills List Api Citizen Bills Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/citizen/pay_bill/{bill_id}:
    post:
      tags:
      - citizen
      summary: Citizen Pay Bill
      operationId: citizen_pay_bill_api_citizen_pay_bill__bill_id__post
      security:
      - HTTPBearer: []
      parameters:
      - name: bill_id
        in: path
        required: true
        schema:
          type: integer
          title: Bill Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitizenBillPayResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/citizen/facilities:
    get:
      tags:
      - citizen
      summary: Citizen Facilities List
      operationId: citizen_facilities_list_api_citizen_facilities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/FacilitySchema'
                  type: array
                type: object
                title: Response Citizen Facilities List Api Citizen Facilities Get
      security:
      - HTTPBearer: []
...
  /api/citizen/bookings:
    get:
      tags:
      - citizen
      summary: Citizen Bookings List
      operationId: citizen_bookings_list_api_citizen_bookings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/FacilityBookingSchema'
                  type: array
                type: object
                title: Response Citizen Bookings List Api Citizen Bookings Get
      security:
      - HTTPBearer: []
...
  /api/commissioner/bookings:
    get:
      tags:
      - commissioner
      summary: Commissioner Bookings List
      operationId: commissioner_bookings_list_api_commissioner_bookings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/FacilityBookingSchema'
                  type: array
                type: object
                title: Response Commissioner Bookings List Api Commissioner Bookings
                  Get
      security:
      - HTTPBearer: []
...
  /api/commissioner/category/{category_id}:
    delete:
      tags:
      - commissioner
      summary: Commissioner Delete Category
      operationId: commissioner_delete_category_api_commissioner_category__category_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Commissioner Delete Category Api Commissioner Category  Category
                  Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/commissioner/bill_types:
    get:
      tags:
      - commissioner
      summary: Commissioner Bill Types List
      operationId: commissioner_bill_types_list_api_commissioner_bill_types_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BillTypeSchema'
                type: array
                title: Response Commissioner Bill Types List Api Commissioner Bill
                  Types Get
      security:
      - HTTPBearer: []
  /api/commissioner/bill_type:
    post:
      tags:
      - commissioner
      summary: Commissioner Create Bill Type
      operationId: commissioner_create_bill_type_api_commissioner_bill_type_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionerBillTypeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionerBillTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/commissioner/bill_type/{bill_type_id}:
    put:
      tags:
      - commissioner
      summary: Commissioner Update Bill Type
      operationId: commissioner_update_bill_type_api_commissioner_bill_type__bill_type_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: bill_type_id
        in: path
        required: true
        schema:
          type: integer
          title: Bill Type Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommissionerBillTypeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionerBillTypeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - commissioner
      summary: Commissioner Delete Bill Type
      operationId: commissioner_delete_bill_type_api_commissioner_bill_type__bill_type_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: bill_type_id
        in: path
        required: true
        schema:
          type: integer
          title: Bill Type Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Commissioner Delete Bill Type Api Commissioner Bill
                  Type  Bill Type Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BillTypeSchema:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: BillTypeSchema
...
    CitizenBillPayResponse:
      properties:
        message:
          type: string
          title: Message
        receipt:
          $ref: '#/components/schemas/CitizenPaymentReceiptSchema'
      type: object
      required:
      - message
      - receipt
      title: CitizenBillPayResponse
    CitizenDashboardResponse:
      properties:
        userName:
          type: string
          title: Username
        totalBillsDue:
          type: integer
          title: Totalbillsdue
        upcomingBookingsCount:
          type: integer
          title: Upcomingbookingscount
        pendingBills:
          items:
            $ref: '#/components/schemas/UtilityBillSchema'
          type: array
          title: Pendingbills
        upcomingBookings:
          items:
            $ref: '#/components/schemas/FacilityBookingSchema'
          type: array
          title: Upcomingbookings
      type: object
      required:
      - userName
      - totalBillsDue
      - upcomingBookingsCount
      - pendingBills
      - upcomingBookings
      title: CitizenDashboardResponse
...
    CitizenPaymentReceiptSchema:
      properties:
        billNumber:
          type: string
          title: Billnumber
        billType:
          type: string
          title: Billtype
        amount:
          type: number
          title: Amount
        paidAt:
          type: string
          title: Paidat
        transactionId:
          type: string
          title: Transactionid
      type: object
      required:
      - billNumber
      - billType
      - amount
      - paidAt
      - transactionId
      title: CitizenPaymentReceiptSchema
...
    CommissionerBillTypeRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      title: CommissionerBillTypeRequest
    CommissionerBillTypeResponse:
      properties:
        message:
          type: string
          title: Message
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - message
      - id
      - name
      title: CommissionerBillTypeResponse
...
    FacilityBookingSchema:
      properties:
        id:
          type: integer
          title: Id
        userId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Userid
        facilityId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Facilityid
        citizenName:
          anyOf:
          - type: string
          - type: 'null'
          title: Citizenname
        facilityName:
          anyOf:
          - type: string
          - type: 'null'
          title: Facilityname
        bookingReference:
          type: string
          title: Bookingreference
        bookedDate:
          type: string
          format: date
          title: Bookeddate
        amountPaid:
          type: number
          title: Amountpaid
        paymentRef:
          anyOf:
          - type: string
          - type: 'null'
          title: Paymentref
        purpose:
          anyOf:
          - type: string
          - type: 'null'
          title: Purpose
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - bookingReference
      - bookedDate
      - amountPaid
      title: FacilityBookingSchema
    FacilitySchema:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        facilityType:
          type: string
          title: Facilitytype
        address:
          type: string
          title: Address
        pincode:
          type: string
          title: Pincode
        pricePerDay:
          type: number
          title: Priceperday
        capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Capacity
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        amenities:
          anyOf:
          - {}
          - type: 'null'
          title: Amenities
        isActive:
          type: boolean
          title: Isactive
      type: object
      required:
      - id
      - name
      - facilityType
      - address
      - pincode
      - pricePerDay
      - isActive
      title: FacilitySchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UtilityBillSchema:
      properties:
        id:
          type: integer
          title: Id
        userId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Userid
        citizenName:
          anyOf:
          - type: string
          - type: 'null'
          title: Citizenname
        billTypeId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Billtypeid
        billType:
          type: string
          title: Billtype
        billNumber:
          type: string
          title: Billnumber
        amount:
          type: number
          title: Amount
        dueDate:
          type: string
          format: date
          title: Duedate
        period:
          anyOf:
          - type: string
          - type: 'null'
          title: Period
        status:
          type: string
          title: Status
        generatedAt:
          type: string
          format: date
          title: Generatedat
        paidAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Paidat
        paymentRef:
          anyOf:
          - type: string
          - type: 'null'
          title: Paymentref
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - billType
      - billNumber
      - amount
      - dueDate
      - status
      - generatedAt
      title: UtilityBillSchema
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer